CI: Remove retries for engine builds
A note on this, I thought the analyzer was flaky, but I think it was instead an incorrect use of source_gen which I've since tracked down and solved. The retries for engine builds aren't working in their current state anyway, and since I don't think we need them anymore, I've oped to remove them instead of trying to fix them.